🐞 fix: 按照实盘修正手续费计算

+++ gmadaptor/config/schema.py
@@ -37,9 +37,9 @@ class Config(object):
         gm_output: Optional[str] = None
 
         class trade_fees:
-            commission: Optional[int] = None
+            commission: Optional[float] = None
 
-            stamp_duty: Optional[int] = None
+            stamp_duty: Optional[float] = None
 
             transfer_fee: Optional[float] = None
